+++ packages/class-core/src/deserializer.ts
@@ -62,10 +62,36 @@ export type Deserializer<T> = (
 
 
 /**
- * A quasi-tuple of the roots (of type {@link INodeBase}) of a model,
+ * A quasi-tuple of the deserialized nodes and roots (both of type {@link INodeBase}) of a model,
  * and its {@link IdMapping} instance.
  */
-export type RootsWithIdMapping = { roots: INodeBase[], idMapping: IdMapping };
+export type DetailedDeserialization = {
+
+    /**
+     * All nodes deserialized by a {@link Deserializer}.
+     */
+    nodes: INodeBase[],
+
+    /**
+     * All root nodes among the deserialized {@link nodes}.
+     * A node is a root node if its serialization did not declare a parent.
+     * *Note* that the parent of a root node might not be actually resolved during this deserialization.
+     */
+    roots: INodeBase[],
+
+    /**
+     * A {@link IdMapping} corresponding to {@link nodes}.
+     */
+    idMapping: IdMapping
+
+};
+
+/**
+ * Legacy alias for {@link DetailedDeserialization}, kept for backward compatibility, and to be removed later.
+ *
+ * @deprecated Use {@link DetailedDeserialization} instead.
+ */
+export type RootsWithIdMapping = DetailedDeserialization;
 
 
 /**
@@ -85,17 +111,17 @@ export type DeserializerConfiguration = {
 
 
 /**
- * @return a {@link Deserializer} function for the given languages (given as {@link ILanguageBase}s) that returns a {@link RootsWithIdMapping}.
+ * @return a {@link Deserializer} function for the given languages (given as {@link ILanguageBase}s) that returns a {@link DetailedDeserialization}.
  * Deprecated:
  * @param languageBases the {@link ILanguageBase}s for (at least) all the languages used in the {@link LionWebJsonChunk} to deserialize, minus LionCore M3 and built-ins.
  * @param receiveDelta an optional {@link DeltaReceiver} that will be injected in all {@link INodeBase nodes} created.
  */
-function nodeBaseDeserializerWithIdMapping(languageBases: ILanguageBase[], receiveDelta?: DeltaReceiver): Deserializer<RootsWithIdMapping>;
+function nodeBaseDetailedDeserializer(languageBases: ILanguageBase[], receiveDelta?: DeltaReceiver): Deserializer<DetailedDeserialization>;
 /**
  * @param configuration a {@link DeserializerConfiguration configuration object} for the deserializer.
  */
-function nodeBaseDeserializerWithIdMapping(configuration: FactoryConfiguration & DeserializerConfiguration): Deserializer<RootsWithIdMapping>;
-function nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration: ILanguageBase[] | (FactoryConfiguration & DeserializerConfiguration), mayBeReceiveDelta?: DeltaReceiver): Deserializer<RootsWithIdMapping> {
+function nodeBaseDetailedDeserializer(configuration: FactoryConfiguration & DeserializerConfiguration): Deserializer<DetailedDeserialization>;
+function nodeBaseDetailedDeserializer(languageBasesOrConfiguration: ILanguageBase[] | (FactoryConfiguration & DeserializerConfiguration), mayBeReceiveDelta?: DeltaReceiver): Deserializer<DetailedDeserialization> {
     const lionWebVersion = (Array.isArray(languageBasesOrConfiguration) ? undefined : languageBasesOrConfiguration.lionWebVersion) ?? LionWebVersions.v2023_1
     const [languageBases, receiveDelta, propertyValueDeserializer, problemReporter] = Array.isArray(languageBasesOrConfiguration)
         ? [languageBasesOrConfiguration, mayBeReceiveDelta, lionWebVersion.builtinsFacade.propertyValueDeserializer, consoleProblemReporter]
@@ -107,7 +133,7 @@ function nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration: ILangua
     return (
         serializationChunk,
         idMapping
-    ): RootsWithIdMapping => {
+    ): DetailedDeserialization => {
 
         const nodesToInstall: NodesToInstall[] = [];
 
@@ -241,6 +267,7 @@ function nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration: ILangua
         }
 
         return {
+            nodes: Object.values(nodesById),
             roots: serializationChunk
                 .nodes
                 .filter(({ parent }) => parent === null)
@@ -251,6 +278,11 @@ function nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration: ILangua
     };
 }
 
+
+/**
+ * Legacy alias for {@link nodeBaseDetailedDeserializer}, kept for backward compatibility, and to be deprecated and removed later.
+ */
+const nodeBaseDeserializerWithIdMapping = nodeBaseDetailedDeserializer
 
 /**
  * @return a {@link Deserializer} function for the languages (given as {@link ILanguageBase}s) that returns the roots (of type {@link INodeBase}) of the deserialized model.
@@ -269,10 +301,10 @@ function nodeBaseDeserializer(languageBasesOrConfiguration: ILanguageBase[] | (F
         idMapping
     ): INodeBase[] =>
         Array.isArray(languageBasesOrConfiguration)
-            ? nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration, receiveDelta)(serializationChunk, idMapping).roots
-            : nodeBaseDeserializerWithIdMapping(languageBasesOrConfiguration)(serializationChunk, idMapping).roots
+            ? nodeBaseDetailedDeserializer(languageBasesOrConfiguration, receiveDelta)(serializationChunk, idMapping).roots
+            : nodeBaseDetailedDeserializer(languageBasesOrConfiguration)(serializationChunk, idMapping).roots
 }
 
 
-export { nodeBaseDeserializerWithIdMapping, nodeBaseDeserializer };
+export { nodeBaseDeserializer, nodeBaseDetailedDeserializer, nodeBaseDeserializerWithIdMapping };